cleaning 3: backend

+++ backend/controllers/photo.controllers.js
@@ -1,22 +1,32 @@
 
-const db = require("./../models");
-const { sequelize, User, Role, Photo, Comment, Like } = require('./../models');
-const Op = db.Sequelize.Op;
+// const db = require("./../models");
+const { User, Photo, Comment, Like } = require('./../models');
+// const Op = db.Sequelize.Op;
 const fs = require("fs");
 
 //-----------------------------------------------------------------------------------------
 
-exports.addPhoto = (req, res)=> {
+exports.addPhoto = async (req, res)=> {
+    try {
+        const image = { 
+          title: req.body.title, 
+          imageUrl:`${req.protocol}://${req.get('host')}/images/photos/${req.file.filename}` 
+        };
 
-    console.log( " REQUETES : ", req)
-    const image = { 
-      title: req.body.title, 
-      imageUrl:`${req.protocol}://${req.get('host')}/images/photos/${req.file.filename}` 
-    };
-    Photo.create( {...image, ownerId: req.userId } )
-    .then(() =>  {
-       res.status(200).json({message: 'Photo Successsfully Posted !'})})
-    .catch( err => res.status(400).json({ message: err.message}) )
+        if(!req.body.title) {
+            const photoName = image.imageUrl.split('/photos/')[1];
+            fs.unlink(`images/photos/${photoName}`, (err) => {
+                if(err) throw err;
+            })
+            throw Error("Title is required !");
+        }
+
+        await Photo.create( {...image, ownerId: req.userId } )
+        return res.status(200).json({message: 'Photo Successsfully Posted !'})
+
+    } catch(err) {
+      res.status(400).json(err.message)
+    }  
 };
 
 //-----------------------------------------------------------------------------------------
@@ -83,6 +93,9 @@ exports.getOnePhoto = (req, res) => {
     // include: [{ model: User, as:'owner'}]
   })
   .then( photo =>  {
+    if (!photo) {
+        return res.status(404).json({ Error: " Photo unknown !"});
+    }
     res.status(200).json(photo)})
   .catch( err => res.status(500).json({ message: err.message || ` Server Error ! Try again Soon `}) )
 }
@@ -90,13 +103,15 @@ exports.getOnePhoto = (req, res) => {
 // -----------------------------------------------------------------------------------------
 
 exports.deleteOnePhoto = (req, res) => {
-  console.log("Hello")
   Photo.findOne( { where: { uuid: req.params.photoUuid } })
   .then( photo =>  {
-    const photoName = photo.imageUrl.split('/photos/')[1];
-    fs.unlink(`images/photos/${photoName}`, (err) => {
-      if(err) throw err;
-    })
+    if (!photo) {
+        return res.status(404).json(" Photo unknown !");
+    }
+    // const photoName = photo.imageUrl.split('/photos/')[1];
+    // fs.unlink(`images/photos/${photoName}`, (err) => {
+    //   if(err) throw err;
+    // })
     Photo.destroy({ where: { uuid: req.params.photoUuid } })
     .then(() => res.status(200).json({ message: ` Photo successfully deleted !`}))
     .catch(err => res.status(403).json({ message: err.message }))
@@ -105,6 +120,7 @@ exports.deleteOnePhoto = (req, res) => {
 }
 
 // -----------------------------------------------------------------------------------------
+
 exports.deleteUserAllPhoto = (req, res) => {
   console.log("Hello");
   Photo.findAll({ where: { ownerId: req.userId } })
